Feat reclean code (#12)
* add lib fn

* remove utils

* move entry to libfn

* remove feature

* split modules into modules

* use fm-coinbase and fm-utxo

* remove comment
