+++ README.md
@@ -1,1 +1,11 @@
 # findorad
+
+## Construction and crates.
+
+- libfindora: Common data struct for findora node, client and wallet.
+- libfn: Functions for wallet and command line client.
+- cli: Provide command line tool called `fn`.
+- findorad: Findora core node called `findorad`.
+- modules:
+  - utxo: Core logic for UTXO.
+  - coinbase: Generate UTXO Outputs based on special rule.
