+++ cli/Cargo.toml
@@ -30,4 +30,6 @@ rand_chacha = "0.2.2"
 bech32 = "0.8.1"
 bip0039 = "0.8.0"
 ed25519-dalek-bip32 = { git = "https://github.com/FindoraNetwork/ed25519-dalek-bip32", branch = "feat-allow-nohardened" }
+libfn = { path = "../libfn" }
+fm-utxo = { path = "../modules/utxo" }
