+++ cli/src/utils/mod.rs
@@ -1,12 +1,13 @@
 use std::collections::BTreeMap;
 
 use abcf_sdk::providers::HttpGetProvider;
-use findorad_lib::utxo_module_rpc::get_owned_outputs;
+use fm_utxo::utxo_module_rpc::get_owned_outputs;
 use libfindora::{transaction::Transaction, utxo::GetOwnedUtxoReq};
 use ruc::*;
 use zei::xfr::{asset_record::open_blind_asset_record, sig::XfrKeyPair, structs::AssetType};
 
-use crate::{config::Config, entry::Entry};
+use crate::{config::Config};
+use libfn::Entry;
 
 pub async fn send_tx(tx: &Transaction) -> Result<()> {
     let provider = abcf_sdk::providers::HttpGetProvider {};
