--- findorad/src/lib.rs
@@ -1,3 +1,0 @@
-#![feature(generic_associated_types)]
-pub mod utxo;
-pub use utxo::utxo_module_rpc;
