+++ findorad/src/main.rs
@@ -1,16 +1,17 @@
 #![feature(generic_associated_types)]
 
-pub mod coinbase;
-pub mod utxo;
+// pub mod coinbase;
+// pub mod utxo;
 
 use std::marker::PhantomData;
-
 use bs3::backend::SledBackend;
 use libfindora::transaction::Transaction;
 use rand_chacha::ChaChaRng;
 use rand_core::SeedableRng;
 use sha3::Sha3_512;
 use zei::setup::PublicParams;
+use fm_utxo::UtxoModule;
+use fm_coinbase::CoinbaseModule;
 
 #[abcf::manager(
     name = "findorad",
@@ -20,20 +21,20 @@ use zei::setup::PublicParams;
     transaction = "Transaction"
 )]
 pub struct Findorad {
-    pub coinbase: coinbase::CoinbaseModule,
-    pub utxo: utxo::UtxoModule,
+    pub coinbase: CoinbaseModule,
+    pub utxo: UtxoModule,
 }
 
 fn main() {
     env_logger::init();
 
-    let coinbase = coinbase::CoinbaseModule::new();
+    let coinbase = CoinbaseModule::new();
 
     let params = PublicParams::default();
 
     let prng = ChaChaRng::from_entropy();
 
-    let utxo = utxo::UtxoModule::new(params, prng);
+    let utxo = UtxoModule::new(params, prng);
 
     let manager = Findorad::<SledBackend>::new(coinbase, utxo);
 
@@ -41,7 +42,7 @@ fn main() {
     let utxo_backend = bs3::backend::sled_db_open(Some("./target/findorad/utxo")).unwrap();
 
     let stateful = abcf::Stateful::<Findorad<SledBackend>> {
-        coinbase: abcf::Stateful::<coinbase::CoinbaseModule<SledBackend>> {
+        coinbase: abcf::Stateful::<CoinbaseModule<SledBackend>> {
             asset_owner: bs3::SnapshotableStorage::new(
                 Default::default(),
                 SledBackend::open_tree(&coinbase_backend, "asset_owner").unwrap(),
@@ -49,7 +50,7 @@ fn main() {
             .unwrap(),
             __marker_s: PhantomData,
         },
-        utxo: abcf::Stateful::<utxo::UtxoModule<SledBackend>> {
+        utxo: abcf::Stateful::<UtxoModule<SledBackend>> {
             output_set: bs3::SnapshotableStorage::new(
                 Default::default(),
                 SledBackend::open_tree(&coinbase_backend, "output_set").unwrap(),
@@ -60,7 +61,7 @@ fn main() {
     };
 
     let stateless = abcf::Stateless::<Findorad<SledBackend>> {
-        coinbase: abcf::Stateless::<coinbase::CoinbaseModule<SledBackend>> {
+        coinbase: abcf::Stateless::<CoinbaseModule<SledBackend>> {
             sl_value: abcf::bs3::SnapshotableStorage::new(
                 Default::default(),
                 SledBackend::open_tree(&utxo_backend, "sl_value").unwrap(),
@@ -68,7 +69,7 @@ fn main() {
             .unwrap(),
             __marker_s: PhantomData,
         },
-        utxo: abcf::Stateless::<utxo::UtxoModule<SledBackend>> {
+        utxo: abcf::Stateless::<UtxoModule<SledBackend>> {
             owned_outputs: abcf::bs3::SnapshotableStorage::new(
                 Default::default(),
                 SledBackend::open_tree(&utxo_backend, "owned_outputs").unwrap(),
