+++ modules/utxo/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "findorad"
+name = "fm-utxo"
 version = "0.1.0"
 edition = "2018"
 
@@ -8,21 +8,11 @@ edition = "2018"
 [dependencies]
 abcf = { git = "ssh://git@github.com/FindoraNetwork/abcf" }
 abcf-node = { git = "ssh://git@github.com/FindoraNetwork/abcf" }
-bs3 = { git = "ssh://git@github.com/FindoraNetwork/bs3", features = ["sled-backend"], tag = "0.1.0" }
-env_logger = "0.9.0"
+abcf-sdk = { git = "ssh://git@github.com/FindoraNetwork/abcf", default-features = false }
 async-trait = "0.1.51"
-sha3 = "0.9.1"
-digest = "0.9.0"
-serde_json = "1.0.68"
-serde = "1.0.130"
-libfindora = { path = "../libfindora" }
+libfindora = { path = "../../libfindora" }
 zei = { git = "ssh://git@github.com/tiannian/zei", branch = "feat-refactor" }
 rand_core = { version = "0.5", default-features = false, features = ["alloc"] }
 rand_chacha = "0.2.2"
 log = "0.4.14"
-fm-utxo = { path = "../modules/utxo" }
-fm-coinbase = { path = "../modules/coinbase" }
-
-# [lib]
-# name = "findorad_lib"
-# path = "src/lib.rs"
+serde_json = "1.0.68"
