+++ modules/utxo/build.rs
@@ -1,0 +1,1 @@
+fn main() {}
