+++ modules/utxo/src/lib.rs
@@ -1,3 +1,5 @@
+#![feature(generic_associated_types)]
+
 use std::convert::TryInto;
 
 use abcf::{
